+++ src/pages/jinminYang/Main/components/FOOTER_INFO_LIST.js
@@ -1,12 +1,13 @@
 export const FOOTER_INFO_LIST = [
-  { id: 1, menu: `도움말` },
-  { id: 2, menu: `홍보 센터` },
-  { id: 3, menu: `API` },
-  { id: 4, menu: `채용 정보` },
-  { id: 5, menu: `개인 정보처리방침` },
-  { id: 6, menu: `약관` },
-  { id: 7, menu: `위치` },
-  { id: 8, menu: `인기 계정` },
-  { id: 9, menu: `해시태그` },
-  { id: 10, menu: `언어` },
+  { id: 1, menu: `소개` },
+  { id: 2, menu: `도움말` },
+  { id: 3, menu: `홍보 센터` },
+  { id: 4, menu: `API` },
+  { id: 5, menu: `채용 정보` },
+  { id: 6, menu: `개인 정보처리방침` },
+  { id: 7, menu: `약관` },
+  { id: 8, menu: `위치` },
+  { id: 9, menu: `인기 계정` },
+  { id: 10, menu: `해시태그` },
+  { id: 11, menu: `언어` },
 ];
