+++ src/pages/jinminYang/Main/components/Feed.scss
@@ -1,0 +1,80 @@
+@import '/src/styles/variables';
+
+.feed {
+  margin-bottom: 40px;
+  border: $borderColor;
+  background-color: white;
+
+  .userInfo {
+    display: flex;
+    justify-content: space-between;
+    padding: 12px;
+
+    .userSimpleProfile {
+      display: flex;
+      justify-content: left;
+
+      .userImg {
+        width: 35px;
+        height: 35px;
+        margin-right: 10px;
+        border: $borderColor;
+        border-radius: 50%;
+      }
+
+      .userName {
+        display: flex;
+        align-items: center;
+      }
+    }
+    .userMenu {
+      display: flex;
+      align-items: center;
+      font-weight: bold;
+    }
+  }
+  .feedImg {
+    img {
+      width: 100%;
+    }
+  }
+  .feedTool {
+    padding: 10px;
+    border: $check-line;
+    img {
+      width: 10px;
+    }
+  }
+  .whoLike {
+    padding: 10px;
+    // border: $check-line;
+  }
+  .replyList {
+    padding: 0 10px;
+  }
+
+  .writeTime {
+    padding: 3px 10px 10px;
+  }
+
+  .replyForm {
+    position: relative;
+
+    input {
+      width: 100%;
+      padding: 12px 10px;
+      border: none;
+      border-top: $borderColor;
+    }
+
+    button {
+      position: absolute;
+      top: 50%;
+      right: 10px;
+      transform: translate(0, -50%);
+      border: none;
+      background-color: white;
+      color: blue;
+    }
+  }
+}
