+++ src/pages/jinminYang/Main/components/components/Reply.scss
@@ -1,3 +1,7 @@
+li {
+  padding: 2px 0px;
+}
+
 .user {
   font-weight: bold;
 }
