+++ src/pages/jinminYang/components/Nav.scss
@@ -2,6 +2,7 @@
 
 .nav {
   position: fixed;
+  z-index: 100;
   top: 0px;
   width: 100%;
   height: 50px;
